Adding more typeDoc
