+++ src/util/logging/LogFactory.ts
@@ -1,9 +1,19 @@
 import { AppConfig } from "../../model/AppConfig"
 import { LogLevel } from "../../model/logging/LogLevel"
 import { ILogger, LogFormat } from "./ILogger"
-import { Logger } from "./Logger"
+import { ConsoleLogger } from "./ConsoleLogger"
 
+/**
+ * Builds implementations of @see{ILogger}.
+ */
 export class LogFactory {
+    /**
+     * Build a new log factory using application config.
+     *
+     * @param appConfig Config object to read logging config from.
+     * @param logLevel (Optional) Lowest level to log, defaults to `info`.
+     * @param logFormat (Optional) Format to output log messages in, defaults to `string`.
+     */
     public constructor(
         appConfig: AppConfig,
         private readonly logLevel: LogLevel = LogLevel.info,
@@ -20,17 +30,34 @@ export class LogFactory {
         }
     }
 
+    /**
+     * Create a new logger.
+     *
+     * @param clazz The enclosing class that will use the new logger.
+     */
     public getLogger(clazz: Function): ILogger {
-        return new Logger(clazz ? clazz.name : "?", this.logLevel, this.logFormat)
+        return new ConsoleLogger(clazz ? clazz.name : "?", this.logLevel, this.logFormat)
     }
 
+    /**
+     * Create a new logger using @see{AppConfig} config defaults.
+     *
+     * @param clazz The enclosing class that will use the new logger.
+     */
     public static getDefaultLogger(clazz: Function) {
         let logFactory = new LogFactory(new AppConfig())
 
         return logFactory.getLogger(clazz)
     }
 
-    public getCustomLogger(clazz: Function, level: LogLevel, format: LogFormat = "string") {
+    /**
+     * Create a new logger using custom log configuration.
+     *
+     * @param clazz The enclosing class that will use the new logger.
+     * @param logLevel (Optional) Lowest level to log, defaults to `info`.
+     * @param logFormat (Optional) Format to output log messages in, defaults to `string`.
+     */
+    public static getCustomLogger(clazz: Function, level: LogLevel = LogLevel.info, format: LogFormat = "string") {
         let logFactory = new LogFactory({
             serverLogger: {
                 format,
